pyproject.toml
[project]
name = "expert_matcher"
version = "0.1.0"
description = "A expert matcher application"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "psycopg>=3.2.5",
    "psycopg-pool>=3.2.5",
    "pydantic>=2.10.6",
    "python-socketio>=5.12.1",
    "python-ulid>=1.1.0",
    "aiohttp>=3.10.11",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/expert_matcher"] 

[dependency-groups]
dev = [
    "black>=24.8.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
]
